[Main] Added rpm release numbers adjustment
This includes:
 * Creation of image from original repository
 * Mounting this iamge to the chroot directory
 * Searching for working qemu binary in the original repository
 * Unpacking this qemu package to the chroot directory
 * Copying and installing rpmrebuild to the chroot directory
 * For each marked package chrooting to that directory, patching the
   package using the installed rpmrebuild, so that to adjust the
   release number to the same number of original package.

This is done because now we are using libzypp as a package manager by
default, and it is more strict in the sense of dependency checking.
Without this patch zypp will fail during the combined image creation
because release numbers of original and marked packages usually differ.
If there is at least one package that depends on the marked one and
requires its particular version and release, the dependcy checking will
fail.

Change-Id: Idf13dbccfe58c3ab2993250f3ab45bbb07653212
Signed-off-by: Ilya Palachev <i.palachev@samsung.com>