[All] added HTTP authentication support
To fix @TC-447 the HTTP authentication support was added, because our
new download servers are available only via authentication.
The user should set "user" and "password" option in config file or in
command line. In the first case the password will be replaced with
string encoded with base64 and bz2 algorithms.

Change-Id: I70b1b3b4f9ece50d048bb91c29bb01eb5149bf4a
Signed-off-by: Ilya Palachev <i.palachev@samsung.com>