[Directory_downloader] Added check of file size
To fix @TC-499 file size check was added. It is doen due to the
possibility of corrupted file download. Now before the downloading all
file sizes are retrieved from server, and after download the resulting
file sizes are compared with them. In case when they are different, file
is redownoaded once more until the resulting size becomes what it must
be.

Change-Id: Ic24673ab67b0848976b4ab6b717925155ed56105
Signed-off-by: Ilya Palachev <i.palachev@samsung.com>